requirements.txt
absl-py==2.3.1
astor==0.8.1
astunparse==1.6.3
beautifulsoup4==4.11.1
boto3==1.37.38
botocore==1.37.38
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.17.1
chardet==4.0.0
charset-normalizer==2.0.4
clip==1.0
colorama==0.4.6
conda==22.9.0
conda-build==3.23.3
conda-package-handling==2.0.2
conda_package_streaming==0.7.0
contourpy==1.1.1
cryptography==39.0.1
cycler==0.12.1
Cython==3.2.4
easydict==1.13
filelock==3.9.0
flatbuffers==25.12.19
flit_core==3.6.0
fonttools==4.57.0
fsspec==2025.3.0
ftfy==6.2.3
gast==0.4.0
glob2==0.7
google-auth==2.47.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.11.0
idna==3.4
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.2
jmespath==1.0.1
jpeg4py==0.1.4
jsonpatch==1.33
jsonpointer==3.0.0
keras==2.13.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
lazy_loader==0.4
libarchive-c==5.1
libclang==18.1.1
lmdb==1.7.5
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.7.5
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mock==4.0.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.3.1
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pandas==2.0.3
pillow==10.4.0
pip==22.3.1
pkginfo==1.8.3
polars==1.8.2
protobuf==4.25.8
psutil==7.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.7
pycosat==0.6.4
pycparser==2.23
pyOpenSSL==23.0.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-pretrained-bert==0.6.2
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.3
regex==2024.11.6
requests==2.28.1
requests-oauthlib==2.0.0
rsa==4.9.1
ruamel-yaml-conda==0.15.100
s3transfer==0.11.5
scikit-image==0.21.0
scipy==1.10.1
setuptools==65.6.3
six==1.17.0
soupsieve==2.3.2.post1
sympy==1.13.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
thop==0.1.1.post2601112201
tifffile==2023.7.10
tikzplotlib==0.10.1
timm==0.5.4
toml==0.10.2
toolz==0.12.0
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tornado==6.4.2
tqdm==4.67.1
triton==3.0.0
typing_extensions==4.5.0
tzdata==2025.3
ultralytics==8.4.7
ultralytics-thop==2.0.18
urllib3==1.26.14
visdom==0.2.4
wcwidth==0.2.14
webcolors==24.8.0
websocket-client==1.8.0
Werkzeug==2.2.2
wheel==0.38.4
wrapt==2.0.1
yacs==0.1.8
zipp==3.20.2
zstandard==0.19.0
